src/covid_project/requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
djangorestframework-xml==2.0.0
